Order transactions latest first
Most of the time, I want to see the latest transaction first, and the oldest transaction last.

Unfortunately (bizarrely?) the Monzo API does not allow for this. What do we think of this as a solution?

Existing behaviour remains default to avoid breaking existing implementations. No tests, as wanted to gauge opinion first...